Backend/requirements.txt
openai
python-dotenv
langchain
langchain-community